Major changes to implement buffer types with focus and keybindings.
The FSM definition that baked in the key binding into dispatched
functions did not seem to be as extensible as a more emacsy
buffer/keybinding approach.  I've implemented modes which simply force
new sets of key-bindings on the *current-buffer*.  Presently, only
the *prompt* buffer has any keybindings, but it should extend pretty
well to support other ways of entering text.